Extract Az Graph ops in separate class Rename GraphOperations to MicrosoftGraphOperations to avoid confusion
