+++ src/api/Endpoints/StorageAccounts.cs
@@ -4,7 +4,6 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using System.Web;
-using Azure.Core;
 using Azure.Identity;
 using Azure.Storage.Files.DataLake;
 using Microsoft.AspNetCore.Http;
@@ -15,7 +14,6 @@ using Microsoft.Extensions.Logging;
 using Microsoft.UsEduCsu.Saas.Services;
 using static Microsoft.UsEduCsu.Saas.FileSystems;
 using Azure.Storage.Files.DataLake.Models;
-using System.Text.Json;
 using System.Globalization;
 
 namespace Microsoft.UsEduCsu.Saas
@@ -51,7 +49,7 @@ namespace Microsoft.UsEduCsu.Saas
 			}
 		}
 
-		internal static List<ContainerDetail> PopulateContainerDetail(string account, string principalId, ILogger log)
+		private static List<ContainerDetail> PopulateContainerDetail(string account, string principalId, ILogger log)
 		{
 			// Get Storage Account Uri
 			var storageUri = SasConfiguration.GetStorageUri(account);
@@ -68,7 +66,7 @@ namespace Microsoft.UsEduCsu.Saas
 				.GetAccessibleContainerDetails(principalId, account);
 
 			// User Operations
-			var graphOps = new GraphOperations(log, ApiCredential);
+			var graphOps = new MicrosoftGraphOperations(log, ApiCredential);
 
 			// Initilize the result
 			ConcurrentBag<ContainerDetail> containerDetails = new();
@@ -81,7 +79,8 @@ namespace Microsoft.UsEduCsu.Saas
 						.Select(l => new { Name = l.Name, Properties = l.Properties })
 						.ToList();
 
-				string accountResourceId = roleOperations.GetAccountResourceId(account);
+				ResourceGraphOperations rgo = new(log, ApiCredential);
+				string accountResourceId = rgo.GetAccountResourceId(account);
 
 				// Build additional details
 				Parallel.ForEach(filesystems, (fs) =>
@@ -103,7 +102,7 @@ namespace Microsoft.UsEduCsu.Saas
 				.ToList();
 		}
 
-		private static ContainerDetail GetContainerDetail(RoleOperations roleOps, GraphOperations graphOps,
+		private static ContainerDetail GetContainerDetail(RoleOperations roleOps, MicrosoftGraphOperations graphOps,
 			string account, string accountResourceId, string container, FileSystemProperties properties,
 			ILogger log)
 		{
@@ -128,7 +127,8 @@ namespace Microsoft.UsEduCsu.Saas
 				{ "Storage Blob Data Reader", 3 } };
 
 			// Determine Access Roles
-			var roles = roleOps.GetStorageDataPlaneRoles(accountResourceId, container);
+			// TODO: Optimization opportunity: Retrieve the role assignments for the account once, and then only the assignments at the container scope
+			var roles = roleOps.GetStorageDataPlaneRoleAssignments(accountResourceId, container);
 			var rbacEntries = roles
 				.Where(r => validTypes.Contains(r.PrincipalType))       // Only display User and Groups (no Service Principals)
 				.Select(r => new StorageRbacEntry()
